requirements.txt
pytest
scikit-learn>=0.23.2
scipy>=1.5.0
numpy==1.19.5
statsmodels>=0.12.0
pyyaml>=5.3.1
tqdm>=4.50.2
nitime==0.9
hyppo>=0.2.1
pyEDM>=1.0.3.2
jpype1>=1.2.0
sktime==0.4.3
dill>=0.3.2
spectral-connectivity>=0.2.4.dev0
umap-learn>=0.4.6
torch>=1.7.0
cdt==0.5.23
oct2py==5.2.0
tslearn==0.5.1
mne==0.22.0
dtw-python==1.1.10
seaborn
